qcom: Use sha256 to build the hashtree in avb image
The hashtree is used in verified boot, and sha256 is more robust against
malicious attacks. Also, sha256 uses the same space as sha1 in the
hashtree. And there isn't much performance regression per
https://b.corp.google.com/issues/156162446#comment18

By putting the config here, we enable sha256 on all targets.
And devices who want to use a different hash algorithm
can override it in it's own board configs.

Bug: 156162446
Test: boot the device and check performance
Change-Id: I9f1d3bcf241bc65adf10376cc5ae7ab1986216fa
Signed-off-by: klozz <carlosj@klozz.dev>